bike rocket damage slightly tweaked++
Bike rocket damage increased to 30.
This lets it kill MLRS in 4 shots instead of 5. Important for the role the bike plays.
Kills Buggy in 4 shots instead of 5 as well. 
This helps to justify bike's cost of $500. Its primary job is to exploit open space with its speed to attack vulnerable units, e.g. go behind enemy lines and quickly kill artillery. 5 shots (3 volleys instead of 2) makes this far weaker in this role, and expensive to take the risk with.