increased speed of TowerMissile; inc. burst delay
Increased speed of TowerMissile. At present it is very slow and not terribly effective vs. air units because they can quickly leave range before getting hit. Speed stats are same for E3 and RocketBike.
Burstdelay increased. This is to avoid wasted missiles on the same target that might get destroyed.
Rangelimit reduced from 40 to 30 to compensate for increased missile speed. 