Sniper range reduced from 7 -> 5.5
Sniper range of 7 seems ridiculous, since it's a one-shot kill, and you can build multiple. (I support building multiple). Surprised I never noticed this issue before.
The range in C&C 95 was 5.5. 
The ScanRadius is set to 5, which is probably why we haven't noticed any OP problems. 
It's also a bit ridiculous since the unit's sight radius is only 5.