nerf GDI turrets
It's really difficult to kill guard towers once they're up. The only reliable ways are to try to take out their power with airstrikes/orcas/etc. or use artillery from a distance. They are supposed to be tough but they are a bit too much.
Guard tower HP should be nerfed because they are wickedly hard to kill. Adv. tower's power should increase because they are annoying when sprawled everywhere and really hard to take out once established.