APC gun damage reduced  vs. heavy armor, infantry
Since the APC's cost was lowered to $600 to deal with orca spam better, its damage vs. heavy armor and infantry is being correspondingly lowered, so that it is not too good against those units (in terms of dollar-value, especially).

Keep in mind that APC is still good at crushing infantry.

Also removed "High" attribute from APC gun (but not from the AA version).