Removed bib from repair bay
Hopefully no one objects to this. This removes the bib from the repair bay.
Rationale: to save space (it's 3x4, which is pretty sizable). It is a building you'll often want to build near the front line, so having it be very large limits whether you can put it on a little rock somewhere. Removing the bib also has the effect of distinguishing it from the starport. It doesn't look bad without the bib.