return orca range to 5
Problem: orcas in small groups will sometimes get in each other's way and be stuck outside of range. The old range was 5, but this was to nerf large groups.

Orcas have been nerfed in several ways. One of the ways was to reduce their range from 5 to 4. However, when the range reduction is coupled with all the other nerfs, it seems a bit too much. Orcas are fairly vulnerable now that air defense has improved.