reduce orca reload time, similar to earlier levels
Orca was over-nerfed with respect to reload time. It was not effective at finishing off vehicles and buildings if it ran out of ammo, because it took way too long to reload.

Its reload time has been set to the old one, but now it reloads 2 pips at a time, so it is slightly slower than originally, but not by much. It's a bit OP when it loads one at a time, it seems.