Nod Artillery buffed. Esp. vs. Heavy armor
If you want to consider increasing the price as a result of this, that might be OK.
But as far as the unit's tactical effectiveness is concerned, I think these changes are necessary, vs. buildings and heavy armor in particular.

In order to keep Arty from being OP, its gun remains inaccurate. It's also slow and vulnerable to exploding.