Reduce DetectCloaked range on towers, HQ
The detect cloaked range is too high, especially for simple guard towers.

The HQ and Advanced Comm. Ctr.'s range of 8 is also absurdly high, and many players don't know about it. The Stealth Tanks ought to be able to sneak through a gap in a base's defenses. Towers should not have unbeatable detection range.