Fix Obelisk ROF
Obelisk ROF wasn't working properly. Its delay between shots was far higher than the 90 intended.
I think the programming for AttackCharge doesn't work as whoever put the data in thought. It works now. 
It now has a ROF of 90 ticks, as intended. For its first shot in a series, the delay is only 50 ticks. 

I think the problem might be that FireDelay and ReloadTime were counted cumulatively.