Minor: adjusting tiberium storage in refinery/silo
It seemed from playing noobs that lack of tiberium storage was an issue for them, because I saw them often building silos, often many of them, and they were usually full. Granted, an experienced player should spend that money faster, but I think it's reasonable to give them $2000 worth of storage in their refinery. 

It seems that how pips are displayed has changed from before, so someone set silos to have 10 pips instead of 12 so that it fits the reticle. That's fine. But maybe it would make sense to just change that to an even $2000 capacity also.