Removed empty top rows from RA base structure footprints
Removed top row of buildings where it only consisted of _ cells and adjusted the art offset instead. This fixes the issue that CenterPosition was located in the middle between the occupied and passable cell for defenses like tesla coils, makes their placement a bit more intuitive and makes setting up proper HitShapes easier.