Fix up the changelog
Just use a dummy changelog entry about the release and fix
a syntax error in the entry footer.

The changelog should not contain the contents of Changelog.md as this
results in an unparsable Debian changelog.  The game's changelog is already
available as CHANGELOG.html in /usr/share/doc/openra.